+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11;
 
-import static org.apache.coyote.http11.Constants.CRLF;
+import static org.apache.coyote.http11.common.Constants.CRLF;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -12,6 +12,13 @@ import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
 import nextstep.jwp.model.User;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.common.HttpMethod;
+import org.apache.coyote.http11.common.HttpStatus;
+import org.apache.coyote.http11.request.RequestBody;
+import org.apache.coyote.http11.request.RequestHeader;
+import org.apache.coyote.http11.request.RequestLine;
+import org.apache.coyote.http11.response.HttpResponseGenerator;
+import org.apache.coyote.http11.response.ResponseEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
