+++ tomcat/src/main/java/org/apache/coyote/http11/exception/Http11Exception.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.exception;
 
 public class Http11Exception extends RuntimeException {
