+++ tomcat/src/main/java/org/apache/coyote/http11/exception/InvalidHttpMethodException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.exception;
 
 public class InvalidHttpMethodException extends Http11Exception {
