+++ tomcat/src/main/java/org/apache/coyote/http11/exception/InvalidRequestLineException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.exception;
 
 public class InvalidRequestLineException extends Http11Exception {
